Switch from PiBakery to pi-maker
So we used PiBakery, but it was hard to customize, and most of the code was
in a big bash script anyway.

Now we have to use the commandline, but there are less steps, and we can build
an image from any platform that Docker supports.

This is a good first step to allowing for self-replication via the raspberry pi.